tidy: Fix implicit float->double conversions.
The clang-tidy "type promotion in math function" check pointed out a
couple of places where floating point numbers are implicitly converted
to double precision numbers.  Replace these calls to math functions
with functions that don't cause the implicit type promotion.  Changes
made by the clang-tidy program.

https://clang.llvm.org/extra/clang-tidy/checks/performance-type-promotion-in-math-fn.html

+++ mythtv/libs/libmythtv/visualisations/goom/mathtools.h
@@ -22,5 +22,5 @@
 #define SINCOS(f,s,c) \
   __asm__ __volatile__ ("fsincos" : "=t" (c), "=u" (s) : "0" (f))
 #else
-#define SINCOS(f,s,c) {(s)=sin(f);(c)=cos(f);}
+#define SINCOS(f,s,c) {(s)=sinf(f);(c)=cosf(f);}
 #endif
